Rename hextobase64 package to set1
Will put all functions in the set in the same package, rather than
creating a new package for each challenge.
