+++ set1/set1_test.go
@@ -1,4 +1,4 @@
-package hextobase64
+package set1
 
 import "testing"
